# Speed up tests with api.yourls.org