Speed up things by saving external HTTP requests

Before:
```
 phpunit --no-coverage --filter HTTP_AYO_Tests tests/tests/http/api-check.php
YOURLS installed, starting PHPUnit

PHPUnit 7.5.20 by Sebastian Bergmann and contributors.

Runtime:       PHP 7.4.3 with Xdebug 2.8.1
Configuration: D:\home\planetozh\ozh.in\phpunit.xml

.............                                                     13 / 13 (100%)

You should really speed up these slow tests (>500ms)...
 1. 934ms to run HTTP_AYO_Tests:test_check_core_version
 2. 818ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #6
 3. 776ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #2
 4. 732ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #5
 5. 705ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #0
 6. 690ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #8
 7. 671ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #1
 8. 663ms to run HTTP_AYO_Tests:test_api_check_in_various_scenario with data set #4


Time: 8.61 seconds, Memory: 12.00 MB

OK (13 tests, 17 assertions)
```

After:
```
$ phpunit --no-coverage --filter HTTP_AYO_Tests tests/tests/http/api-check.php
YOURLS installed, starting PHPUnit

PHPUnit 7.5.20 by Sebastian Bergmann and contributors.

Runtime:       PHP 7.4.3 with Xdebug 2.8.1
Configuration: D:\home\planetozh\ozh.in\phpunit.xml

.............                                                     13 / 13 (100%)

Time: 2.64 seconds, Memory: 12.00 MB

OK (13 tests, 17 assertions)
```